% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_croissance.R
\name{dataframe_croissance}
\alias{dataframe_croissance}
\title{Poids
calculate the weight of the beast for one day in particular you must loop on jour and phase}
\usage{
dataframe_croissance(njour, tableGmq, tablePhases, PI, Alim, Qte, tableAlim,
  tableViande)
}
\arguments{
\item{njour}{the day when you want the weight}

\item{tableGmq}{dataframe that contain the gmq}

\item{tablePhases}{dataframe that contain the Phase}

\item{PI}{initial weight}
}
\description{
Poids
calculate the weight of the beast for one day in particular you must loop on jour and phase
}
