% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegime.R
\name{plotRegime}
\alias{plotRegime}
\title{plotRegime
plot the variables of the tableCroiss dataframe}
\usage{
plotRegime(tableCroiss, var)
}
\arguments{
\item{tableCroiss}{dataframe containing the growing variables weigth and price of food}

\item{var}{variable to plot}
}
\description{
plotRegime
plot the variables of the tableCroiss dataframe
}
