% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimModule.R
\name{regim}
\alias{regim}
\title{regim
Server part of the module that create the regim and visualisation}
\usage{
regim(input, output, session, nphase)
}
\arguments{
\item{nphase}{numeric, number of phases}

\item{nalim}{numeric, number of aliment}
}
\description{
regim
Server part of the module that create the regim and visualisation
}
