% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_alim.R
\name{rhandsontable_alim}
\alias{rhandsontable_alim}
\title{table_alim
Creat the handsontable widget that contain the food price infos}
\usage{
rhandsontable_alim(n_alim, nom_alim)
}
\description{
table_alim
Creat the handsontable widget that contain the food price infos
}
