% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gmq.R
\name{rhandsontable_GMQ}
\alias{rhandsontable_GMQ}
\title{table GMQ
Creat the handsontable widget that contain the GMQ infos}
\usage{
rhandsontable_GMQ(n_alim, n_phase)
}
\arguments{
\item{n_alim}{numeric, number of aliment}

\item{n_phase}{numeric, number of phase}
}
\description{
table GMQ
Creat the handsontable widget that contain the GMQ infos
}
