% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_phase.R
\name{rhandsontable_phase}
\alias{rhandsontable_phase}
\title{table GMQ
Creat the handsontable widget that contain the GMQ infos}
\usage{
rhandsontable_phase(n_phase)
}
\arguments{
\item{n_phase}{numeric, number of phase}
}
\description{
table GMQ
Creat the handsontable widget that contain the GMQ infos
}
