% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_viande.R
\name{rhandsontable_viande}
\alias{rhandsontable_viande}
\title{table_viande
Creat the handsontable widget that contain the beast price infos}
\usage{
rhandsontable_viande(nphase)
}
\description{
table_viande
Creat the handsontable widget that contain the beast price infos
}
